Создан макет для MainActivity с содержанием тестового задания, а также контактной информацией разработчике и кнопками для проверки тестового задания. Настроено открытие фрагмента с информацией о еде, а также возвращением обратно в MainActivity с информацией о задании.
