+++ app/src/main/java/com/food/meal/order/foodview/view/activity/MainActivity.kt
@@ -1,19 +1,13 @@
 package com.food.meal.order.foodview.view.activity
 
-import android.graphics.Typeface
 import android.os.Bundle
+import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.FragmentManager
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import com.food.meal.order.foodview.R
 import com.food.meal.order.foodview.databinding.ActivityMainBinding
 import com.food.meal.order.foodview.navigator.BackButtonListener
-import com.food.meal.order.foodview.utils.FONT_INTER
-import com.food.meal.order.foodview.utils.FONT_ROBOTO
 import com.food.meal.order.foodview.utils.MAIN_ACTIVITY_SCOPE
-import com.food.meal.order.foodview.view.foodviewfragment.adapters.FoodListRecyclerAdapter
-import com.food.meal.order.foodview.view.foodviewfragment.adapters.KindFoodListRecyclerAdapter
 import com.github.terrakok.cicerone.NavigatorHolder
 import com.github.terrakok.cicerone.androidx.AppNavigator
 import org.koin.core.qualifier.named
@@ -33,20 +27,6 @@ class MainActivity: AppCompatActivity(), FragmentManager.OnBackStackChangedListe
     // Навигация
     private val navigator = AppNavigator(this@MainActivity, R.id.fragments_container)
     private val navigatorHolder: NavigatorHolder = KoinJavaComponent.getKoin().get()
-
-    // Временные данные для проверки работоспособности макета
-    val kindFoodList: List<String> = listOf("Пицца", "Комбо", "Десерты", "Напитки",
-        "Пицца", "Комбо", "Десерты", "Напитки", "Пицца", "Комбо", "Десерты", "Напитки")
-    val foodList: List<String> = listOf("Пицца", "Изюм", "Мандарины", "Баклажаны", "Пицца", "Изюм",
-        "Мандарины", "Баклажаны", "Пицца", "Изюм", "Мандарины", "Баклажаны", "Пицца", "Изюм",
-        "Мандарины", "Баклажаны", "Пицца", "Изюм", "Мандарины", "Баклажаны", "Пицца", "Изюм",
-        "Мандарины", "Баклажаны", "Пицца", "Изюм", "Мандарины", "Баклажаны", "Пицца", "Изюм",
-        "Мандарины", "Баклажаны", "Пицца", "Изюм", "Мандарины", "Баклажаны", "Пицца", "Изюм",
-        "Мандарины", "Баклажаны")
-    // Горизонтальный список видов еды
-    lateinit var kindFoodListRecyclerView: RecyclerView
-    // Вертикальный список еды
-    lateinit var foodListRecyclerView: RecyclerView
     //endregion
 
     override fun onDestroy() {
@@ -66,31 +46,10 @@ class MainActivity: AppCompatActivity(), FragmentManager.OnBackStackChangedListe
             // Установка текущего экрана приложения
             navigatorHolder.setNavigator(navigator)
         }
-
-        // Установка списка видов еды
-        kindFoodListRecyclerView = binding.kindFoodList
-        kindFoodListRecyclerView.layoutManager = LinearLayoutManager(
-            this, LinearLayoutManager.HORIZONTAL, false)
-        kindFoodListRecyclerView.adapter = KindFoodListRecyclerAdapter(kindFoodList)
-        // Установка списка еды
-        foodListRecyclerView = binding.foodList
-        foodListRecyclerView.layoutManager = LinearLayoutManager(
-            this, LinearLayoutManager.VERTICAL, false)
-        foodListRecyclerView.adapter = FoodListRecyclerAdapter(foodList)
-
-        // Установка шрифтов элементам макета
-        setFontsToElements()
-
+        // Инициализация кнопок для проверки тестового задания
+        initCheckButtons()
         // Отображение содержимого окна
         setContentView(binding.root)
-    }
-
-    // Установка шрифтов элементам
-    private fun setFontsToElements() {
-        binding.cityTitle.typeface = Typeface.createFromAsset(assets, FONT_ROBOTO)
-        binding.menuTitle.typeface = Typeface.createFromAsset(assets, FONT_INTER)
-        binding.profileTitle.typeface = Typeface.createFromAsset(assets, FONT_INTER)
-        binding.basketTitle.typeface = Typeface.createFromAsset(assets, FONT_INTER)
     }
 
     // Инициализация ViewModel
@@ -112,11 +71,14 @@ class MainActivity: AppCompatActivity(), FragmentManager.OnBackStackChangedListe
         super.onPause()
     }
     override fun onBackPressed() {
+        // Отображение текста с тестовым заданием:
+        binding.taskDescriptionScroll.visibility = View.VISIBLE
         supportFragmentManager.fragments.forEach {
             if (it is BackButtonListener && it.backPressed()) {
                 return
             }
         }
+        // Закрыите приложения
         viewModel.router.exit()
     }
     override fun onBackStackChanged() {
@@ -124,4 +86,16 @@ class MainActivity: AppCompatActivity(), FragmentManager.OnBackStackChangedListe
         if (supportFragmentManager.backStackEntryCount == 0) finish()
     }
     //endregion
+
+    // Инициализация кнопок для проверки тестового задания
+    private fun initCheckButtons() {
+        binding.moveToFoodViewFragmentButtonTop.setOnClickListener {
+            viewModel.router.navigateTo(viewModel.screens.foodViewScreen())
+            binding.taskDescriptionScroll.visibility = View.GONE
+        }
+        binding.moveToFoodViewFragmentButtonBottom.setOnClickListener {
+            viewModel.router.navigateTo(viewModel.screens.foodViewScreen())
+            binding.taskDescriptionScroll.visibility = View.GONE
+        }
+    }
 }
