+++ app/src/main/java/com/food/meal/order/foodview/view/foodviewfragment/FoodViewFragment.kt
@@ -1,9 +1,18 @@
 package com.food.meal.order.foodview.view.foodviewfragment
 
 import android.content.Context
+import android.graphics.Typeface
+import android.os.Bundle
+import android.view.View
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import com.food.meal.order.foodview.databinding.FragmentFoodViewBinding
 import com.food.meal.order.foodview.model.base.BaseFragment
+import com.food.meal.order.foodview.utils.FONT_INTER
+import com.food.meal.order.foodview.utils.FONT_ROBOTO
 import com.food.meal.order.foodview.utils.FOOD_VIEW_FRAGMENT_SCOPE
+import com.food.meal.order.foodview.view.foodviewfragment.adapters.FoodListRecyclerAdapter
+import com.food.meal.order.foodview.view.foodviewfragment.adapters.KindFoodListRecyclerAdapter
 import org.koin.core.qualifier.named
 import org.koin.core.scope.Scope
 import org.koin.java.KoinJavaComponent
@@ -19,6 +28,20 @@ class FoodViewFragment:
     companion object {
         fun newInstance(): FoodViewFragment = FoodViewFragment()
     }
+
+    // Временные данные для проверки работоспособности макета
+    val kindFoodList: List<String> = listOf("Пицца", "Комбо", "Десерты", "Напитки",
+        "Пицца", "Комбо", "Десерты", "Напитки", "Пицца", "Комбо", "Десерты", "Напитки")
+    val foodList: List<String> = listOf("Пицца", "Изюм", "Мандарины", "Баклажаны", "Пицца", "Изюм",
+        "Мандарины", "Баклажаны", "Пицца", "Изюм", "Мандарины", "Баклажаны", "Пицца", "Изюм",
+        "Мандарины", "Баклажаны", "Пицца", "Изюм", "Мандарины", "Баклажаны", "Пицца", "Изюм",
+        "Мандарины", "Баклажаны", "Пицца", "Изюм", "Мандарины", "Баклажаны", "Пицца", "Изюм",
+        "Мандарины", "Баклажаны", "Пицца", "Изюм", "Мандарины", "Баклажаны", "Пицца", "Изюм",
+        "Мандарины", "Баклажаны")
+    // Горизонтальный список видов еды
+    lateinit var kindFoodListRecyclerView: RecyclerView
+    // Вертикальный список еды
+    lateinit var foodListRecyclerView: RecyclerView
     //endregion
 
     /** Работа со Scope */ //region
@@ -35,4 +58,31 @@ class FoodViewFragment:
         super.onDetach()
     }
     //endregion
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        // Установка списка видов еды
+        kindFoodListRecyclerView = binding.kindFoodList
+        kindFoodListRecyclerView.layoutManager = LinearLayoutManager(
+            requireContext(), LinearLayoutManager.HORIZONTAL, false)
+        kindFoodListRecyclerView.adapter = KindFoodListRecyclerAdapter(kindFoodList)
+        // Установка списка еды
+        foodListRecyclerView = binding.foodList
+        foodListRecyclerView.layoutManager = LinearLayoutManager(
+            requireContext(), LinearLayoutManager.VERTICAL, false)
+        foodListRecyclerView.adapter = FoodListRecyclerAdapter(foodList)
+
+        // Установка шрифтов элементам макета
+        setFontsToElements()
+    }
+
+    // Установка шрифтов элементам
+    private fun setFontsToElements() {
+        binding.cityTitle.typeface = Typeface.createFromAsset(requireContext().assets, FONT_ROBOTO)
+        binding.menuTitle.typeface = Typeface.createFromAsset(requireContext().assets, FONT_INTER)
+        binding.profileTitle.typeface =
+            Typeface.createFromAsset(requireContext().assets, FONT_INTER)
+        binding.basketTitle.typeface = Typeface.createFromAsset(requireContext().assets, FONT_INTER)
+    }
 }
