+++ app/src/main/res/values/dimens.xml
@@ -39,4 +39,21 @@
     <dimen name="food_item_container_margin_top">16dp</dimen>
     <dimen name="food_item_container_margin_bottom">16dp</dimen>
     <dimen name="kind_food_container_padding_bottom">16dp</dimen>
+    <dimen name="author_info_text_size">20sp</dimen>
+    <dimen name="author_contact_text_size">20sp</dimen>
+    <dimen name="todo_main_title_margin_top">8dp</dimen>
+    <dimen name="todo_main_title_text_size">20sp</dimen>
+    <dimen name="todo_main_text_part_one_text_size">20sp</dimen>
+    <dimen name="todo_main_text_part_two_text_size">20sp</dimen>
+    <dimen name="todo_option_title_text_size">20sp</dimen>
+    <dimen name="todo_option_text_text_size">20sp</dimen>
+    <dimen name="result_format_title_margin_top">8dp</dimen>
+    <dimen name="result_format_title_text_size">20sp</dimen>
+    <dimen name="result_format_text_text_size">20sp</dimen>
+    <dimen name="limitations_title_margin_top">8dp</dimen>
+    <dimen name="limitations_title_text_size">20sp</dimen>
+    <dimen name="limitations_text_text_size">20sp</dimen>
+    <dimen name="task_description_scroll_margin">8dp</dimen>
+    <dimen name="move_to_food_view_fragment_button_bottom_margin_top">8dp</dimen>
+    <dimen name="move_to_food_view_fragment_button_bottom_margin_bottom">20dp</dimen>
 </resources>
