+++ app/src/main/res/values/strings.xml
@@ -7,4 +7,16 @@
     <string name="food_name_tools_text">Еда</string>
     <string name="food_ingredients_text_sample">Баварские колбаски, ветчина, пикантная пипперони, острая чоризо, томатны соус</string>
     <string name="food_price_button_text_sample">от 300 р</string>
+    <string name="author_info_text">Тестовое задание\nХахалина Андрея Владимировича></string>
+    <string name="author_contact_text"><![CDATA[контактная информация\n(копирование текста - longClick):\n+79055364506, avkhakhalin@mail.ru\nрепозиторий: https://github.com/AVKhakhalin/FoodView\nрезюме: https://hh.ru/search/vacancy?text=%D0%A0%D0%B0%D0%B7%D1%80%D0%B0%D0%B1%D0%BE%D1%82%D1%87%D0%B8%D0%BA+android&from=suggest_post&salary=&clusters=true&area=1&ored_clusters=true&label=not_from_agency&enable_snippets=true)\n]]></string>
+    <string name="todo_main_title_text">Что надо было сделать:</string>
+    <string name="todo_main_text_part_one_text">● главный экран приложения доставки еды, дизайн по ссылке (можно открыть в браузере по клику):\nhttps://www.figma.com/file/8FvAWXCD2oD9oSDHx9xFfU/%D0%A2%D0%B5%D1%81%D1%82%D0%BE%D0%B2%D0%BE%D0%B5-%D0%B7%D0%B0%D0%B4%D0%B0%D0%BD%D0%B8%D0%B5-Android?node-id=0%3A1;\n</string>
+    <string name="todo_main_text_part_two_text">● в баннеры можно захардкодить любые фото;\n● основная задача - сделать идентичную планку с категориями и блок меню(Подсказка. Для реализации можно использовать CollapsingToolbarLayout);\n● планка с категориями при скролле должна прилипать к верхнему бару(для примера можно посмотреть приложение Додо Пицца);\n● в качестве API использовать любой открытый источник подходящий под текущие нужды.\n</string>
+    <string name="todo_option_title_text">Опционально:\n</string>
+    <string name="todo_option_text_text">●\tOffline-режим: т.е. в случае, если нет доступа к сети, показывать последние загруженные данные(и ленту, и детали).</string>
+    <string name="result_format_title_text">Результат работ на отправку:</string>
+    <string name="result_format_text_text">● apk-debug файл.\n● ссылка на репозиторий ()\n● ссылка на резюме.</string>
+    <string name="limitations_title_text">Ограничения на стек технологий:</string>
+    <string name="limitations_text_text">● MVVM;\n● Clean Architecture;\n● Остальное на ваше усмотрение.</string>
+    <string name="move_to_food_view_fragment_button_text">Проверить задание</string>
 </resources>
